requirements.txt
# TensorFlow 2.10 with DirectML for AMD GPU support
tensorflow-cpu==2.10.0
tensorflow-directml-plugin

# Image processing libraries
imageio>=2.25.0
scikit-image>=0.19.0
Pillow>=9.3.0
opencv-python>=4.7.0.0

# Numerical computing
numpy>=1.21.0,<1.24.0  # TF 2.10 requires numpy<1.24
matplotlib>=3.5.0

# Optional: for better performance
tqdm>=4.64.0
